Move common entity code to HWAMStoveEntity in entity.py Derive entity classes from HWAMStoveEntity.
